Add files via upload
This repository contains a clone of the Cult.fit website created using HTML, CSS, and a touch of JavaScript. The project is part of a learning journey to enhance web development skills. Cult.fit is a fitness and wellness platform, and this clone aims to replicate some of its features and design.